requirements.txt
torch==2.0.1
torchvision==0.15.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
mlflow==2.8.1
dvc==3.30.1
pytest==7.4.3
requests==2.31.0
prometheus-client==0.19.0
kaggle==1.5.16
tqdm==4.66.1
